Move rendering logic to formatters.html
